+++ scripts/configs/types.ts
@@ -57,11 +57,14 @@ export interface PoolConfig {
     PROOF_OF_RESERVE?: Hex;
     PROOF_OF_RESERVE_AGGREGATOR?: Hex;
     DELEGATION_AWARE_A_TOKEN_IMPL_REV_1?: Hex;
-    STATIC_A_TOKEN_FACTORY?: Hex; // deprecated
+    LEGACY_STATIC_A_TOKEN_FACTORY?: Hex; // deprecated
+    LEGACY_STATIC_A_TOKEN_ORACLE?: Hex; // deprecated
     STATA_FACTORY?: Hex;
-    STATIC_A_TOKEN_ORACLE?: Hex;
     CAPS_PLUS_RISK_STEWARD?: Hex;
     RISK_STEWARD?: Hex;
+    EDGE_RISK_STEWARD?: Hex;
+    EDGE_STEWARD_INJECTOR?: Hex;
+    RISK_ORACLE?: Hex;
     FREEZING_STEWARD?: Hex;
     UI_GHO_DATA_PROVIDER?: Hex;
     RATES_FACTORY?: Hex;
@@ -106,8 +109,6 @@ export interface NetworkAddresses<T extends Record<string, AddressInfo> = {}> {
   name: string;
   chainId: number;
   addresses: {
-    // https://github.com/bgd-labs/aave-paraswap-claimer
-    PARASWAP_FEE_CLAIMER?: Hex;
     TRANSPARENT_PROXY_FACTORY?: Hex;
     PROXY_ADMIN?: Hex;
     CREATE_3_FACTORY?: Hex;